// ui/src/app/api-gen/model/execCommand.ts
/**
 * odo dev
 * API interface for \'odo dev\'
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExecCommand { 
    component: string;
    commandLine: string;
    workingDir: string;
    hotReloadCapable: boolean;
}